feat: Auth utility functions
Create utility functions for signIn, signOut, and createUser that
handle the respective authentication actions for firebase.

`signIn` accepts an email and password; then, returns a promise for a
Firebase UserCredential object if sign in is a success. Rejects with
an error if sign in fails.

`signOut` signs out the current user and returns a promise.

`createUser` accepts and email and password; then returns a promise for
a Firebase UserCredential object if the user account is created and
signed in successfully. Rejects with an error if sign in fails.

- All Firebase calls are mocked.
- Create dependency on Firebase.